# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer & Career Coach API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserTier = typeof UserTier[keyof typeof UserTier];


export const UserTier = {
  free: 'free',
  pro: 'pro',
} as const;

export interface User {
  id: string;
  clerkId: string;
  email: string;
  /** @nullable */
  name?: string | null;
  tier: UserTier;
  /** @nullable */
  stripeCustomerId?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  createdAt: string;
}

export interface UsageStatus {
  used: number;
  limit: number;
  remaining: number;
  isPro: boolean;
}

export interface AnalysisSuggestion {
  issue: string;
  before: string;
  after: string;
}

export interface AnalysisResult {
  score: number;
  summary: string;
  strengths: string[];
  weaknesses: string[];
  suggestions: AnalysisSuggestion[];
  atsKeywords: string[];
}

export interface Analysis {
  id: string;
  userId: string;
  resumeText: string;
  jobDescription: string;
  score: number;
  result: AnalysisResult;
  createdAt: string;
}

export interface AnalysisInput {
  resumeText: string;
  jobDescription: string;
}

export interface AnalysisStats {
  total: number;
  thisMonth: number;
  /** @nullable */
  avgScore: number | null;
  tier: string;
}

export interface CheckoutSession {
  url: string;
}

export interface PortalSession {
  url: string;
}