# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Resume Analyzer & Career Coach API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User management
  - name: analyses
    description: Resume analysis operations
  - name: billing
    description: Stripe billing operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get or create current user record
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /users/me/usage:
    get:
      operationId: getMyUsage
      tags: [users]
      summary: Get today's usage count for the current user
      responses:
        "200":
          description: Usage data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageStatus"
        "401":
          description: Unauthorized

  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List past analyses for the current user
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"
        "401":
          description: Unauthorized
    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Run AI analysis on a resume + job description and save it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisInput"
      responses:
        "201":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "429":
          description: Rate limit exceeded

  /analyses/stats:
    get:
      operationId: getAnalysisStats
      tags: [analyses]
      summary: Get aggregate stats for the current user's analyses
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisStats"
        "401":
          description: Unauthorized

  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get a single analysis by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteAnalysis
      tags: [analyses]
      summary: Delete an analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /billing/create-checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Create a Stripe Checkout session for the Pro plan
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "401":
          description: Unauthorized

  /billing/portal:
    post:
      operationId: createPortalSession
      tags: [billing]
      summary: Create a Stripe Customer Portal session
      responses:
        "200":
          description: Portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalSession"
        "401":
          description: Unauthorized

  /billing/webhook:
    post:
      operationId: stripeWebhook
      tags: [billing]
      summary: Stripe webhook endpoint (raw body, signature verified)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /clerk/webhook:
    post:
      operationId: clerkWebhook
      tags: [users]
      summary: Clerk webhook for user.created events
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    User:
      type: object
      required: [id, clerkId, email, tier, createdAt]
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        tier:
          type: string
          enum: [free, pro]
        stripeCustomerId:
          type: ["string", "null"]
        stripeSubscriptionId:
          type: ["string", "null"]
        createdAt:
          type: string

    UsageStatus:
      type: object
      required: [used, limit, remaining, isPro]
      properties:
        used:
          type: integer
        limit:
          type: integer
        remaining:
          type: integer
        isPro:
          type: boolean

    AnalysisSuggestion:
      type: object
      required: [issue, before, after]
      properties:
        issue:
          type: string
        before:
          type: string
        after:
          type: string

    AnalysisResult:
      type: object
      required: [score, summary, strengths, weaknesses, suggestions, atsKeywords]
      properties:
        score:
          type: integer
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisSuggestion"
        atsKeywords:
          type: array
          items:
            type: string

    Analysis:
      type: object
      required: [id, userId, resumeText, jobDescription, score, result, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        resumeText:
          type: string
        jobDescription:
          type: string
        score:
          type: integer
        result:
          $ref: "#/components/schemas/AnalysisResult"
        createdAt:
          type: string

    AnalysisInput:
      type: object
      required: [resumeText, jobDescription]
      properties:
        resumeText:
          type: string
        jobDescription:
          type: string

    AnalysisStats:
      type: object
      required: [total, thisMonth, avgScore, tier]
      properties:
        total:
          type: integer
        thisMonth:
          type: integer
        avgScore:
          type: ["integer", "null"]
        tier:
          type: string

    CheckoutSession:
      type: object
      required: [url]
      properties:
        url:
          type: string

    PortalSession:
      type: object
      required: [url]
      properties:
        url:
          type: string