# packages/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer & Career Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResult } from './analysisResult';

export interface Analysis {
  id: string;
  userId: string;
  resumeText: string;
  jobDescription: string;
  score: number;
  result: AnalysisResult;
  createdAt: string;
}