# packages/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer & Career Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisSuggestion } from './analysisSuggestion';

export interface AnalysisResult {
  score: number;
  summary: string;
  strengths: string[];
  weaknesses: string[];
  suggestions: AnalysisSuggestion[];
  atsKeywords: string[];
}