# packages/api-zod/src/generated/types/analysisSuggestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer & Career Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisSuggestion {
  issue: string;
  before: string;
  after: string;
}