# packages/api-zod/src/generated/types/usageStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer & Career Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface UsageStatus {
  used: number;
  limit: number;
  remaining: number;
  isPro: boolean;
}