# packages/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer & Career Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { UserTier } from './userTier';

export interface User {
  id: string;
  clerkId: string;
  email: string;
  /** @nullable */
  name?: string | null;
  tier: UserTier;
  /** @nullable */
  stripeCustomerId?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  createdAt: string;
}